# Reduce heap spikes