--- a/testdata/Makefile
+++ b/testdata/Makefile
@@ -4,7 +4,7 @@ CC = tcc
 CXX = clang++
 CXXFLAGS ?= -O2
 
-all: clang_hello tcc_hello rust_hello rust_hello_stripped
+all: clang_hello tcc_hello rust_hello rust_hello_stripped go_hello
 
 clang_hello: clang_hello.cpp
 	${CXX} ${CXXFLAGS} $< -o $@
@@ -23,5 +23,8 @@ rust_hello_stripped: rust_hello.rs
 	strip $@
 	chmod -x $@
 
+go_hello: main.go
+	GOOS=linux GOARCH=amd64 go build -o go_hello main.go
+
 clean:
-	rm -f clang_hello tcc_hello rust_hello
+	rm -f clang_hello tcc_hello rust_hello go_hello