Previously we read entire ELF sections into memory all at once and then
search the byte slice. This was problematic because depending on the
given ELF binary we load an unpredictable amount of heap.

This patch makes use of the StreamReader introduced in the first commit to read
the ELF sections in chunks. (think of `tail <file> | grep <xyz>`)

Not only did this remove all heap spikes.

To test this, we used a very large Rust binary that was not stripped.
Before these changes we observed a heap of ~1GB.

```
$ /usr/bin/time -l -h -p go test -run TestRustXYZ
PASS
ok      github.com/xyproto/ainur        1.697s
          1096564736  maximum resident set size
          ...
```

After these changes, we observed a heap of ~80mb (which is likely just
due to operating system accounting, nothing was found in heap profiling
data anymore).

```
$ /usr/bin/time -l -h -p go test -run TestRustXYZ
PASS
ok      github.com/xyproto/ainur        0.202s
            84819968  maximum resident set size
            ...
```

Lastly, not only did this remove heap spikes, we could also observe a
massive improvement in runtime speed:

```
$ benchstat old.txt new.txt
name                old time/op    new time/op    delta
RustXYZ-10     1.54s ± 5%     0.01s ± 1%   -99.61%  (p=0.008 n=5+5)

name                old alloc/op   new alloc/op   delta
RustXYZ-10    2.81GB ± 0%    0.00GB ± 1%  -100.00%  (p=0.008 n=5+5)

name                old allocs/op  new allocs/op  delta
RustXYZ-10     24.4k ± 0%      0.4k ± 0%   -98.44%  (p=0.000 n=4+5)
```

Likely due to removed allocations.

Co-Authored-By: Thor Hansen <me@thor-hansen.com>