@xyproto 2023-03-10
That's really cool! Thanks for the PR. I'll give it a spin and then merge it.